Fixed the sign of the angle in image.rotate() method
The image.rotate() method used in Operations.py was using a positive angle for performing clockwise rotation and a negative angle for anticlockwise rotation, which was opposite with the behavior of the image.rotate() function that is when this function is given positive angle of rotation then it rotates in anticlockwise direction and for negative angle of rotation it rotates in clockwise direction. This commit corrects the sign of the angle of rotation provided inside the image.rotate() method to ensure the rotation is performed correctly according to the desired clockwise and anticlockwise directions.